% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvi.R
\name{compute_kwon}
\alias{compute_kwon}
\title{Calculate the Kwon index}
\usage{
compute_kwon(x, u, v, m = 2)
}
\arguments{
\item{x}{A numeric matrix of observations.}

\item{u}{Membership matrix.}

\item{v}{Cluster centers.}

\item{m}{Fuzziness parameter.}
}
\value{
A numeric value representing the Kwon index.
}
\description{
Compute the Kwon cluster validity index from data, membership matrix,
and cluster centers.
}
