% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_limit.r
\name{factor_limit}
\alias{factor_limit}
\title{factor_limit function}
\usage{
factor_limit(t, c, lim)
}
\arguments{
\item{t}{Table of data}

\item{c}{Selected column to test}

\item{lim}{Limit to number of levels}
}
\description{
This function builds a ggplot2 function to plot data
}
\examples{
factor_limit()
}
\keyword{factor}
\keyword{limits}
