% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscales.R
\name{mModels}
\alias{mModels}
\title{Get Measurement Models for the questionnaire used}
\usage{
mModels(data, scales, run.models, ...)
}
\arguments{
\item{data}{The data frame containing only the variables that are used for the analyses}

\item{scales}{A character vector of scales used.}

\item{run.models}{Logical. Should the models be run or should they be returnes as character output? Default is to \code{FALSE}, which returns only the definition of the measurement models in lavaan syntax.}

\item{...}{If \code{run.models = TRUE}, a measurement model for each scale is fitted, using \code{\link[lavaan]{cfa}}. These \code{...} arguments are passed on directly to this function.}
}
\value{
Depending on \code{run.models}, this function returns either a list of character vectors, each containint the measurement model in \code{lavaan} syntax or the fitted model.
}
\description{
Get Measurement Models for the questionnaire used
}
\examples{
scales <-  c("E = 1r, 6, 11R, 16",
             "A = 2r, 7, 12r, 17r",
             "C = 3, 8r, 13, 18",
             "N = 4, 9r, 14, 19",
             "O = 5, 10, 15, 20 21")
data <- bigfive[grep(x = names(bigfive), pattern = "BFI_")]
mModels(data, scales, run.models = FALSE, std.lv = TRUE)
}

