% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Visualize spatial cluster labels}
\usage{
plot_clusters(
  dat_loc,
  cluster,
  title = "Cluster",
  palette = NULL,
  point_size = 1.45,
  point_alpha = 1,
  legend_nrow = 2,
  legend_position = "bottom",
  strip_text_size = 16,
  legend_text_size = 12,
  legend_title_size = 12,
  ...
)
}
\arguments{
\item{dat_loc}{A numeric matrix of dimension \eqn{n \times \rho}, where
\eqn{n} is the number of observations and \eqn{\rho = 2} corresponds to
two spatial coordinates.}

\item{cluster}{A vector of length \eqn{n} giving cluster labels.}

\item{title}{A character string specifying the facet title.}

\item{palette}{An optional character vector of colors. If \code{NULL}, ggplot
default discrete colors are used.}

\item{point_size}{A numeric value specifying the point size.}

\item{point_alpha}{A numeric value specifying the point alpha.}

\item{legend_nrow}{An integer specifying the number of rows in the legend.}

\item{legend_position}{A character string specifying the legend position.}

\item{strip_text_size}{A numeric value specifying the facet strip text size.}

\item{legend_text_size}{A numeric value specifying the legend text size.}

\item{legend_title_size}{A numeric value specifying the legend title size.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Visualize spatial cluster labels
}
