% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Visualize uncertainty on spatial coordinates}
\usage{
plot_uncertainty(
  dat_loc,
  uncertainty,
  mode = c("continuous", "category"),
  title = "Uncertainty",
  point_size = 1.45,
  point_alpha = 1,
  tile_alpha = 1,
  add_tile = TRUE,
  relabel = TRUE,
  legend_position = "bottom",
  strip_text_size = 16,
  legend_text_size = 12,
  legend_title_size = 12,
  legend_nrow = 4,
  ...
)
}
\arguments{
\item{dat_loc}{A numeric matrix of dimension \eqn{n \times \rho}, where
\eqn{n} is the number of observations and \eqn{\rho = 2} corresponds to
two spatial coordinates.}

\item{uncertainty}{A numeric vector of length \eqn{n} giving uncertainty values.}

\item{mode}{A character string specifying the visualization mode.
Must be one of \code{"continuous"} or \code{"category"}.}

\item{title}{A character string specifying the facet title.}

\item{point_size}{A numeric value specifying the point size.}

\item{point_alpha}{A numeric value specifying the point alpha.}

\item{tile_alpha}{A numeric value specifying the tile alpha.}

\item{add_tile}{A logical value indicating whether to add \code{geom_tile()}.}

\item{relabel}{A logical value indicating whether to relabel category levels as
"Certain", "Uncertain (moderate)", "Uncertain (high)", and
"Uncertain (extreme)" when \code{mode = "category"}.}

\item{legend_position}{A character string specifying the legend position.}

\item{strip_text_size}{A numeric value specifying the facet strip text size.}

\item{legend_text_size}{A numeric value specifying the legend text size.}

\item{legend_title_size}{A numeric value specifying the legend title size.}

\item{legend_nrow}{An integer specifying the number of rows in the legend when
\code{mode = "category"}.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Uncertainty can be displayed either as a continuous variable or as
categorized levels based on mean and standard deviation thresholds.
}
