% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmclust.R
\name{print.fcmclust}
\alias{print.fcmclust}
\title{Print an fcmclust object}
\usage{
\method{print}{fcmclust}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fcmclust"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
The input object, invisibly.
}
\description{
Print an fcmclust object
}
