% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcmclust_auto.R
\name{print.fcmclust_auto}
\alias{print.fcmclust_auto}
\title{Print an fcmclust_auto object}
\usage{
\method{print}{fcmclust_auto}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fcmclust_auto"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
The input object, invisibly.
}
\description{
Print an fcmclust_auto object
}
