% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscales.R
\name{recodeQuick}
\alias{recodeQuick}
\title{recodeQuick to recode questionnaire data}
\usage{
recodeQuick(data, recode, valid.values = NULL)
}
\arguments{
\item{data}{A data frame that contains only the values of the data frame.}

\item{recode}{A numeric vector giving the poisitions of the items to recode or a character vector giving the names of the variables to recode.}

\item{valid.values}{A vector containing the valid values of the questionnaire, i. e. the ones that should be recoded. Can be useful if missings are recoded as \code{-99}.}
}
\value{
The data frame with the recoded items.
}
\description{
Questionnaire data usually contain only values ranging from 1 to 4 or 1 to 5. This function takes as input the items of a data frame that should be recoded (as numeric position or as names) and recodes these items. Unless missings are scored as \code{NA} and not, for example, as \code{-99}, the function will just reverse the values of the items.
}
\examples{
## eample 1
data <- data.frame(a = c(1, 2, -99), b = c(3, 4, NA))
recodeQuick(data, recode = "a", valid.values = c(1:4))

## example 2
data <- bigfive[grep(x = names(bigfive), pattern = "BFI_")]
data <- data[1:10, 1:10]
data[1, 6] <- -99 ## here, a missing is coded as -99.
recode <- c(6, 10)
recodeQuick(data, recode = c(6, 10), valid.values = c(1:4))
}

