% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollPlot.r
\name{rollPlot}
\alias{rollPlot}
\title{rollPlot rolling plotting function}
\usage{
rollPlot(t, x, y, ylab = "", xlab = "", w = 400, s = 80, f = "mean",
  scale = "linear", feature = "data points", data = F,
  samples = "Samples", split = F, sd = F, cols = vector(), facet = NA,
  splitscale = "fixed")
}
\arguments{
\item{t}{Table of data.}

\item{x}{Variable to be binned on x-axis.}

\item{y}{Vector of variables to plot on y-axis.}

\item{ylab}{Label for Y-axis. Defaults to "variable".}

\item{xlab}{Label for X-axis.}

\item{w}{Window size (number of values per bin). Defaults to 400.}

\item{s}{Step size (number of values to move along for each bin). Defaults to 80.}

\item{f}{Function to apply to y values for each bin ("mean","median","sum"). Defaults to "mean".}

\item{scale}{Scale the x-axis ("linear"=linear spacing of points based value of bin or "bins"=equal spacing between all bins). Defaults to "linear".}

\item{feature}{Name of features being binned. Defaults to "data points".}

\item{data}{Returns a table in long form if set to True.}

\item{samples}{Set the legend name. Defaults to "Samples".}

\item{split}{Plot samples separately.}

\item{sd}{Plot standard deviation as ribbon when plotting mean.}

\item{facet}{Variable to facet by.}

\item{splitscale}{Scales for split plots "fixed"(default),"free",free_x","free_y".}
}
\description{
This function orders and bins a variable in to groups defined by a window 
and step size then computes rolling averages for multipe variables across each bin.
}
\examples{
rollPlot()
}
\keyword{average}
\keyword{bin}
\keyword{plots.}
