% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuest.r
\name{runQuest}
\alias{runQuest}
\title{runQuest function}
\usage{
runQuest()
}
\arguments{
\item{t}{Table of data}

\item{c}{Selected column to test}

\item{lim}{Limit to number of levels}
}
\description{
This function builds a ggplot2 function to plot data
}
\keyword{factor}
\keyword{limits}
\keyword{runQuest()}
