% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscales.R
\name{scale2list}
\alias{scale2list}
\title{Transform a scale into a readable list}
\usage{
scale2list(scale)
}
\arguments{
\item{scale}{A description of a scale in the form of \code{scale <- c(name = 1, 2, 3, 4r, 5)}. The scale is defined using \code{=}, the items by their position in the data frame or questionnaire (this should correspond!).}
}
\value{
Returns a list that list that contains four entries: The input used (\code{original.scale}), the derived name of the scale (\code{name}), the items that constitute the scale \code{items}, the items that need to be recoded (\code{recode})
}
\description{
This function turns a scale in the form of \code{A = 1, 5, 10R, 15r} into a list which of \enumerate{
 \item the scale used
 \item the name of the scale
 \item the items to use for the scale (by position in data frame or questionnaire)
 \item the items that need to be recoded.
}
}
\details{
Items that need to be recoded are indicated using an \code{R} or \code{r} behind the number of the item.
}
\examples{
scales = c("E = 1r, 6, 11R, 16",
           "A = 2r, 7, 12r, 17r",
           "C = 3, 8r, 13, 18",
           "N = , 9r, 14, 19",
           "O = 4, 10RR, 15, 20 21")
scale2list(scales[1])
}

