% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiler.r
\name{tiler}
\alias{tiler}
\title{tiler function}
\usage{
tiler(t, x, xl = F, xs = 1, y, yl = F, ys = 1, z, zl = F, zs = 1,
  bin = 1, zrange = NA, xrange = NA, yrange = NA, func = "median",
  col = NA, smooth = NA)
}
\arguments{
\item{t}{Table of data}

\item{x, y, z}{Variables for each dimension}

\item{xl, yl, zl}{Log the variable first. Defaults to F.}

\item{xs, ys, zs}{Scale the variable first. Defaults to 1.}

\item{bin}{Size of each tile}

\item{zrange}{Range of the colour scale.}

\item{xrange}{Range displayed on x-axis.}

\item{yrange}{Range displayed on y-axis.}

\item{func}{Function applied to tiled z variables ("median","mean","sum","count"). Defaults to median.}

\item{smooth}{Add a smoothing line, options are lm,glm,gam,loess,rlm. Off by default.}

\item{cols}{Vector of colours for colour range. Defaults to colorRamps::matlab.like2(10)}
}
\description{
This function plots 3 variables on an x,y and coloured z scale. Values on the x and y scale are tiled
and the colour of the tile is determined by the function applied to z (mean,median,count etc.).
}
\examples{
tiler()
}
\keyword{plots}
\keyword{tile}

