% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvi.R
\name{compute_xb}
\alias{compute_xb}
\title{Calculate the XB index}
\usage{
compute_xb(x, u, v, m = 2)
}
\arguments{
\item{x}{A numeric matrix of observations.}

\item{u}{Membership matrix.}

\item{v}{Cluster centers.}

\item{m}{Fuzziness parameter.}
}
\value{
A numeric value representing the XB index.
}
\description{
Compute the Xie-Beni index from data, membership matrix, and cluster centers.
}
