% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_builder.r
\name{ggplot_builder}
\alias{ggplot_builder}
\title{ggplot_builder function}
\usage{
ggplot_builder(d, x, y, z, geom = "point", facet = NA, smooth = NA,
  xlim = NA, ylim = NA, xrotate = 0, colour = NA, fill = NA,
  bar.position = "stack", bins = 0, outliers = T, enable.plotly = F,
  theme = "grey", logx = F, logy = F)
}
\arguments{
\item{d}{Table of data}

\item{x, y, z}{Variables for each dimension}

\item{geom}{Select a ggplot2 geometry (currently point,line,histogram,bar,boxplot,violin)}

\item{facet}{Facet plot by values in a column}

\item{smooth}{Add a smooth line to point plots (gam,lm,loess,rlm,glm,auto)}

\item{xlim}{Range displayed on x-axis}

\item{ylim}{Range displayed on y-axis}

\item{xrotate}{Angle to rotate x-axis labels (90=vertical)}

\item{colour}{A variable to colour by}

\item{fill}{A variable to fill by}

\item{bar.position}{Position of bars in a bar plot (stack,dodge,fill)}

\item{bins}{Number of bins to use in a binned plot (histogram)}

\item{outliers}{Set outliers=F to remove outliers from boxplot}

\item{enable.plotly}{convert to interactive Plotly plot}

\item{logx, logy}{Log the variable first. Defaults to F.}

\item{smooth}{Add a smoothing line, options are lm,glm,gam,loess,rlm. Off by default.}
}
\description{
This function builds a ggplot2 function to plot data
}
\examples{
ggplot_builder()
}
\keyword{plots}
\keyword{tile}

