% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_builder.r
\name{ggplot_builder}
\alias{ggplot_builder}
\title{ggplot_builder function}
\usage{
ggplot_builder(d, x, y = NA, z = NA, geom = "point", facet = NA,
  smooth = NA, xlim = NA, ylim = NA, xrotate = 0, colour = NA,
  fill = NA, bar.position = "stack", binwidth = 0, bins = 0,
  outliers = T, varwidth = F, enable.plotly = F, theme = "grey",
  logx = F, logy = F, man_colour = NA, man_fill = NA,
  gradient = "default", gradient.steps = 10, colourset = "default",
  coord_flip = F, cut_method = "number", cut.n = 10, factorlim = 50,
  stat.method = "count", stat.func = "mean")
}
\arguments{
\item{d}{Table of data}

\item{x, y, z}{Variables for each dimension}

\item{geom}{Select a ggplot2 geometry (currently point,line,histogram,bar,boxplot,violin)}

\item{facet}{Facet plot by values in a column}

\item{smooth}{Add a smooth line to point plots (gam,lm,loess,rlm,glm,auto)}

\item{xlim}{Range displayed on x-axis}

\item{ylim}{Range displayed on y-axis}

\item{xrotate}{Angle to rotate x-axis labels (90=vertical)}

\item{colour}{A variable to colour by}

\item{fill}{A variable to fill by}

\item{bar.position}{Position of bars in a bar plot (stack,dodge,fill)}

\item{binwidth}{Size of binwidth in binned plots (histogram)}

\item{bins}{Add a stat_bin with this number of bins}

\item{outliers}{Set outliers=F to remove outliers from boxplot}

\item{varwidth}{Set varwidth=T to plot boxplots with variable width based on dataset size}

\item{enable.plotly}{convert to interactive Plotly plot}

\item{theme}{Set ggplot theme (grey,bw,dark,light,void,linedraw,minimal,classsic)}

\item{logx, logy}{Log the variable first. Defaults to F.}

\item{man_colour}{Select a solid colour}

\item{man_fill}{Select a solid fill colour}

\item{gradient}{Select gradient colour scheme (default,Matlab)}

\item{gradient.steps}{Set outliers=F to remove outliers from boxplot}

\item{colourset}{Select colour scheme (default,Set1,Set2,Set3,Spectral)}

\item{coord_flip}{Flip the x and y axes (default=F)}

\item{cut_method}{Select method for binning continuous X axis in boxplots (number,interval,width see cut_interval etc.)}

\item{cut.n}{Binning number applied to cut_method}

\item{factorlim}{Set maximum levels allowed to use factors for plotting (default=50)}

\item{stat.method}{Set stat method for barplots (count,identity,summary) (default=bin)}

\item{stat.func}{Set summary function for stat.method="summary" (default=mean)}
}
\description{
This function builds a ggplot2 function to plot data
}
\examples{
ggplot_builder()
}
\keyword{builder}
\keyword{ggplot}
\keyword{wrapper}

