% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Visualize spatial cluster labels}
\usage{
plot_clusters(
  dat_loc,
  cluster,
  title = "Cluster",
  palette = NULL,
  point_size = 1.45,
  point_alpha = 1,
  legend_nrow = 2,
  legend_position = "bottom",
  strip_text_size = 16,
  legend_text_size = 12,
  legend_title_size = 12,
  ...
)
}
\arguments{
\item{dat_loc}{A numeric matrix or data frame of spatial coordinates
with two columns.}

\item{cluster}{A vector of cluster labels of length n.}

\item{title}{Character; facet title.}

\item{palette}{Optional character vector of colors. If NULL, ggplot default
discrete colors are used.}

\item{point_size}{Numeric; point size.}

\item{point_alpha}{Numeric; point alpha.}

\item{legend_nrow}{Integer; number of rows in legend.}

\item{legend_position}{Character; legend position.}

\item{strip_text_size}{Numeric; facet strip text size.}

\item{legend_text_size}{Numeric; legend text size.}

\item{legend_title_size}{Numeric; legend title size.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Visualize spatial cluster labels
}
