% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUEST-package.R
\docType{package}
\name{QUEST-package}
\alias{QUEST}
\alias{QUEST-package}
\title{QUEST: Quantifying Uncertainty for Spatial Domain Detection in Spatial Transcriptomics}
\description{
QUEST is a computational method for quantifying uncertainty in spatial domain detection using fuzzy clustering. It provides probabilistic domain assignments and location-level uncertainty estimates, enabling refinement of domain detection, identification of ambiguous regions such as domain boundaries and heterogeneous subdomains, and improvement of downstream analyses by focusing on high-confidence locations.
}
\author{
\strong{Maintainer}: Yanlin Tong \email{zoetong@umich.edu}

}
\keyword{internal}
