% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuest.r
\name{runQuest}
\alias{runQuest}
\title{Run quest app function}
\usage{
runQuest()
}
\description{
Run quest app function
}
\examples{
This function runs the quest application.
quest()
}
\keyword{quest}
\keyword{shiny}

